SOTA+ submission: PR#198 base + TTT + RoPE50K + EMA + curriculum
Rebuild from the proven #1 submission (PR #198, 1.1326 BPB) and stack
four untried improvements:

- RoPE base 50K (smoother position interpolation at seq2048)
- LAWA-EMA replacing periodic SWA (continuous exponential moving average)
- Context-length curriculum (seq1024 early for 60% more steps, seq2048 late)
- Full-model SGD test-time training (1 epoch, lr=3e-4, on val data)

Architecture: 11L 512d MLP3x SmearGate BigramHash OrthoInit WD=0.04
Artifact: ~15.7MB (int6+zstd-22), 26.8M params, FA3 with SDPA fallback
Pending 8xH100 run. Target: sub-1.13 BPB.

Made-with: Cursor
