--- records/track_10min_16mb/2026-03-20_AllInOne_SmearGate_Int6QAT_SlidingWindow/submission.json
@@ -1,12 +1,0 @@
-{
-  "author": "Jaishy Shivashish",
-  "github_id": "0xjaishy",
-  "name": "SmearGate + OrthoInit + Int6 STE QAT + MLP 3x + Bigram Hash + Sliding Window",
-  "blurb": "All-in-one submission stacking every proven technique: SmearGate bigram embedding blending, 4096-bucket bigram hash embeddings, MLP 3x expansion (hidden=1536), orthogonal weight init, STE int6 QAT, fp16 tied embedding passthrough, int6 per-row quantization with zstd-22, sliding window eval stride=64, Muon with decoupled weight decay 0.01 and momentum 0.99, tuned LR schedule (warmdown=3000, matrix_lr=0.02).",
-  "date": "2026-03-20",
-  "val_loss": 0.0,
-  "val_bpb": 0.0,
-  "bytes_total": 0,
-  "bytes_code": 53137,
-  "note": "val_bpb and bytes_total to be filled after 8xH100 run"
-}
